Further model conversion fixes.

+++ Source/ModelConverter/Mesh.h
@@ -7,7 +7,7 @@ struct VertexElement;
 struct Mesh
 {
     std::string materialName;
-    std::vector<uint16_t> faceIndices;
+    std::vector<uint32_t> faceIndices;
     std::vector<std::vector<Vector4>> vertexStreams[14];
     std::vector<VertexElement> vertexElements;
     std::vector<uint16_t> nodeIndices;
